move MessageID and Macro enums to new package
