+++ src/test/java/com/winterhavenmc/util/messagebuilder/YamlLanguageHandlerTests.java
@@ -19,6 +19,7 @@ package com.winterhavenmc.util.messagebuilder;
 
 import be.seeseemelk.mockbukkit.MockBukkit;
 import be.seeseemelk.mockbukkit.ServerMock;
+import com.winterhavenmc.util.messagebuilder.messages.MessageId;
 import org.bukkit.World;
 import org.junit.jupiter.api.*;
